requirements.txt
flask
flask-cors
psycopg2-binary
langchain
langchain-community
langchain-google-genai
pgvector
python-dotenv
langchain-classic
langchain-core
sqlglot
langchain-postgres